+++ gst/datarepo/gstdatareposink.h
@@ -43,20 +43,20 @@ struct _GstDataRepoSink
 
   GstCaps *fixed_caps;            /**< to get meta info */
   JsonObject *json_object;        /**< JSON object */
-  JsonArray *sample_offset_array;   /**< offset array of sample */
-  JsonArray *tensor_size_array;    /**< size array of flexible tensor */
+  JsonArray *sample_offset_array; /**< offset array of sample */
+  JsonArray *tensor_size_array;   /**< size array of flexible tensor */
   JsonArray *tensor_count_array;  /**< array for the number of cumulative tensors */
-  guint cumulative_tensors;    /**< the number of cumulated tensors */
+  guint cumulative_tensors;       /**< the number of cumulated tensors */
 
   gboolean is_static_tensors;
   gint fd;                        /**< open file descriptor*/
   GstDataRepoDataType data_type;  /**< data type */
   gint total_samples;             /**< The number of total samples, in the case of multi-files, it is used as an index. */
   guint64 fd_offset;              /**< offset of fd */
-  guint sample_size;             /**< size of one sample */
+  gsize sample_size;              /**< size of one sample */
 
   /* property */
-  gchar *filename;    /**< filename */
+  gchar *filename;      /**< filename */
   gchar *json_filename; /**< "JSON file path to store the meta information */
 };
